Release

+++ app/GraphQL/Mutations/CommentMutation.php
@@ -1,15 +1,26 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace App\GraphQL\Mutations;
 
 use App\Models\OrderComment;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Str;
+use App\Traits\TeamContextTrait;
 use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 final readonly class CommentMutation
 {
+    use TeamContextTrait;
+
+    private function isMyComment(string $commentId): bool
+    {
+        $user = auth()->user();
+        $comment = OrderComment::find($commentId);
+
+        return $user && $user->id === $comment->user_id;
+    }
+
     /**
      * Return a value for the field.
      *
@@ -22,47 +33,44 @@ final readonly class CommentMutation
     public function updateVisibility(null $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): mixed
     {
         // TODO implement the resolver
-        //$user = auth()->user();
-        //$phone = $args['phone'];
 
-        return OrderComment::updateVisibility($args['commentId'], $args['isPublic']);
+        if ($this->isMyComment($args['commentId'])) {
+            return OrderComment::updateVisibility($args['commentId'], $args['isPublic']);
+        }
+
+        return null;
     }
 
     public function updateText(null $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): mixed
     {
         // TODO implement the resolver
-        //$user = auth()->user();
-        //$phone = $args['phone'];
+        if ($this->isMyComment($args['commentId'])) {
+            return OrderComment::updateText($args['commentId'], $args['text']);
+        }
 
-        return OrderComment::updateText($args['commentId'], $args['text']);
+        return null;
     }
 
     public function delete(null $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): mixed
     {
         // TODO implement the resolver
-        //$user = auth()->user();
-        //$phone = $args['phone'];
+        if ($this->isMyComment($args['commentId'])) {
+            return OrderComment::destroy($args['commentId']);
+        }
 
-        return OrderComment::destroy($args['commentId']);
+        return null;
     }
 
     public function create(null $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): mixed
     {
-        // TODO implement the resolver
-        //$user = auth()->user();
-        //$phone = $args['phone'];
-        $team = DB::table('teams')->first()->id;
-        $user = DB::table('users')->first()->id;
-        //dd($args);
+        $team_id = $this->getTeamIdFromContext($context);
 
         return OrderComment::create([
-            'id' => (string) Str::ulid(),
             'order_id' => $args['orderId'],
-            'team_id' => $team,
+            'team_id' => $team_id,
             'comment' => strip_tags($args['comment']),
-            'user_id' => $user,
+            'user_id' => auth()->user()->id,
             'is_public' => $args['isPublic'],
         ]);
     }
-
 }
